Update Stepper styling and test
