+++ src/lib/index.ts
@@ -19,7 +19,7 @@ export { default as Footer } from './Footer.svelte';
 export { default as Jumbotron } from './jumbotron/Jumbotron.svelte';
 export { default as Loader } from './Loader.svelte';
 export { default as Modal } from './Modal.svelte';
-export { default as Stepper } from './Stepper.svelte';
+export { default as Stepper } from './stepper/Stepper.svelte';
 export { default as TitleDescription } from './TitleDescription.svelte';
 export { default as Tooltip } from './Tooltip.svelte';
 export { default as Typography } from './Typography.svelte';
